+++ packages/next/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @spree/next
 
+## 0.6.5
+
+### Patch Changes
+
+- Rename `per_page` pagination parameter to `limit`
+
 ## 0.6.4
 
 ### Patch Changes